+++ src/plsravg_unif.jl
@@ -1,15 +1,15 @@
-struct PlsrAvgUnif
+struct PlsravgUnif
     fm::Plsr
     nlv
 end
 
-function plsr_avg_unif(X, Y, weights = ones(nro(X)); nlv,
+function plsravg_unif(X, Y, weights = ones(nro(X)); nlv,
         scal = false)
-    plsr_avg_unif!(copy(ensure_mat(X)), copy(ensure_mat(Y)), weights; nlv = nlv,
+    plsravg_unif!(copy(ensure_mat(X)), copy(ensure_mat(Y)), weights; nlv = nlv,
         scal = scal)
 end
 
-function plsr_avg_unif!(X::Matrix, Y::Matrix, weights = ones(nro(X)); nlv,
+function plsravg_unif!(X::Matrix, Y::Matrix, weights = ones(nro(X)); nlv,
         scal = false)
     X = ensure_mat(X)
     n, p = size(X)
@@ -17,10 +17,10 @@ function plsr_avg_unif!(X::Matrix, Y::Matrix, weights = ones(nro(X)); nlv,
     nlv = (min(minimum(nlv), n, p):min(maximum(nlv), n, p))
     nlvmax = maximum(nlv)     
     fm = plskern!(X, Y, weights; nlv = nlvmax, scal = scal)
-    PlsrAvgUnif(fm, nlv)
+    PlsravgUnif(fm, nlv)
 end
 
-function predict(object::PlsrAvgUnif, X)
+function predict(object::PlsravgUnif, X)
     nlv = object.nlv
     le_nlv = length(nlv)
     zpred = predict(object.fm, X; nlv = nlv).pred